@@ -11,51 +11,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.3",
-    "@angular/cdk": "^7.1.1",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/fire": "^5.1.1",
-    "@angular/forms": "~7.1.0",
-    "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
-    "@fortawesome/angular-fontawesome": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.10",
-    "@fortawesome/free-solid-svg-icons": "^5.6.1",
-    "add": "^2.0.6",
+    "@angular/animations": "^7.2.13",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^7.2.13",
+    "@angular/compiler": "^7.2.13",
+    "@angular/core": "^7.2.13",
+    "@angular/fire": "^5.1.2",
+    "@angular/forms": "^7.2.13",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^7.2.13",
+    "@angular/platform-browser-dynamic": "^7.2.13",
+    "@angular/router": "^7.2.13",
     "angular-font-awesome": "^3.1.2",
-    "bootstrap": "^4.1.3",
-    "core-js": "^2.5.4",
-    "firebase": "^5.7.0",
-    "firebase-admin": "^6.4.0",
-    "firebase-functions": "^2.1.0",
+    "core-js": "^2.6.5",
+    "firebase": "^5.9.4",
     "font-awesome": "^4.7.0",
+    "ngx-ckeditor": "^0.4.0",
+    "ngx-image-cropper": "^1.3.9",
+    "ngx-spinner": "^7.1.4",
     "pdfjs-dist": "^2.0.943",
     "rxjs": "~6.3.3",
     "tslib": "^1.9.0",
-    "yarn": "^1.12.3",
-    "zone.js": "~0.8.26"
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.11.3",
-    "@angular/cli": "^7.1.3",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "^0.13.8",
+    "@angular/cli": "^7.3.8",
+    "@angular/compiler-cli": "^7.2.13",
+    "@angular/language-service": "^7.2.13",
+    "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
     "codelyzer": "~4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
+    "karma": "^3.1.4",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "protractor": "^5.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
     "typescript": "~3.1.6"
