@@ -4,55 +4,60 @@ import {NgModule} from '@angular/core';
 import {AppComponent} from './app.component';
 import {AngularFireModule} from '@angular/fire';
 import {environment} from '../environments/environment';
-import {AngularFireAuthModule} from '@angular/fire/auth';
 import {AppRoutingModule} from './app-routing.module';
 
-import {AuthService} from './core/auth.service';
-import {UserLoginComponent} from './users/user-login/user-login.component';
-import {UserProfileComponent} from './users/user-profile/user-profile.component';
-import {AngularFirestoreModule} from '@angular/fire/firestore';
+import {AuthService} from './users/auth.service';
 import {AngularFontAwesomeModule} from 'angular-font-awesome';
-import {PdfService} from './core/pdf.service';
+import {PdfService} from './upload/pdf.service';
 import {CoursesComponent} from './ui/courses/courses.component';
 import {CourseComponent} from './ui/course/course.component';
 import {QuestionComponent} from './ui/question/question.component';
 import {ExamComponent} from './ui/exam/exam.component';
 import {HeaderComponent} from './ui/header/header.component';
-import {BreadcrumbsComponent} from './ui/breadcrumbs/breadcrumbs.component';
-import {UploadComponent} from './ui/upload/upload.component';
-import {SemesterPipe} from './core/semester.pipe';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {MatInputModule} from '@angular/material';
 import {MatCheckboxModule} from '@angular/material/checkbox';
-import {AngularFireStorageModule} from '@angular/fire/storage';
+import {CoreModule} from './core/core.module';
+import {FormsModule} from '@angular/forms';
+import {FacultiesComponent} from './ui/faculties/faculties.component';
+import {FacultyComponent} from './ui/faculty/faculty.component';
+import {MsGraphService} from './users/msgraph.service';
+import {NgxSpinnerModule} from 'ngx-spinner';
+import {SolutionComponent} from './ui/solution/solution.component';
+import {UsersModule} from './users/users.module';
+import {UploadModule} from './upload/upload.module';
+import {AdminModule} from './admin/admin.module';
+import {DiscussionsModule} from './discussions/discussions.module';
 
 @NgModule({
+  imports: [
+    BrowserModule,
+    AngularFireModule.initializeApp(environment.firebase),
+    AppRoutingModule,
+    AngularFontAwesomeModule,
+    CoreModule,
+    AdminModule,
+    DiscussionsModule,
+    UploadModule,
+    UsersModule,
+    FormsModule,
+    BrowserAnimationsModule,
+    MatInputModule,
+    MatCheckboxModule,
+    NgxSpinnerModule
+  ],
   declarations: [
     AppComponent,
-    UserLoginComponent,
-    UserProfileComponent,
     HeaderComponent,
-    BreadcrumbsComponent,
     CoursesComponent,
     CourseComponent,
     QuestionComponent,
     ExamComponent,
-    UploadComponent,
-    SemesterPipe
+    SolutionComponent,
+    FacultiesComponent,
+    FacultyComponent
   ],
-  imports: [
-    BrowserModule,
-    AngularFireModule.initializeApp(environment.firebase),
-    AngularFireAuthModule,
-    AngularFirestoreModule,
-    AngularFireStorageModule,
-    AppRoutingModule,
-    AngularFontAwesomeModule,
-    BrowserAnimationsModule,
-    MatInputModule,
-    MatCheckboxModule
-  ],
-  providers: [AuthService, PdfService],
+  providers: [AuthService, PdfService, MsGraphService],
   bootstrap: [AppComponent]
 })
 export class AppModule {
