@@ -1,10 +1,9 @@
-import { Pipe, PipeTransform } from '@angular/core';
+import {Pipe, PipeTransform} from '@angular/core';
 
 @Pipe({
   name: 'semester'
 })
 export class SemesterPipe implements PipeTransform {
-
   transform(value: number, args?: any): string {
     switch (value) {
       case 1: return 'Winter';
@@ -13,5 +12,4 @@ export class SemesterPipe implements PipeTransform {
       default: return 'Unknown semester';
     }
   }
-
 }
