@@ -1,6 +1,6 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { CourseComponent } from './course.component';
+import {CourseComponent} from './course.component';
 
 describe('CourseComponent', () => {
   let component: CourseComponent;
