@@ -1,6 +1,6 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { CoursesComponent } from './courses.component';
+import {CoursesComponent} from './courses.component';
 
 describe('CoursesComponent', () => {
   let component: CoursesComponent;
