@@ -1,6 +1,6 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { ExamComponent } from './exam.component';
+import {ExamComponent} from './exam.component';
 
 describe('ExamComponent', () => {
   let component: ExamComponent;
