@@ -1,8 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
 import {DbService} from '../../core/db.service';
-import {Question, QuestionId} from '../../core/entities/question';
-import {Exam} from '../../core/entities/exam';
+import {QuestionId} from '../../entities/question';
+import {Exam} from '../../entities/exam';
+import {AuthService} from '../../users/auth.service';
+import {Course} from '../../entities/course';
 
 @Component({
   selector: 'app-exam',
@@ -13,12 +15,17 @@ export class ExamComponent implements OnInit {
 
   private courseId: number;
   private examId: string;
-  public questions: QuestionId[];ß
+  public questions: QuestionId[];
   public exam: Exam;
+  course: Course;
 
-  constructor(private route: ActivatedRoute, private db: DbService) {
+  adminAccess: boolean;
+
+  constructor(private route: ActivatedRoute, private db: DbService, private auth: AuthService) {
     this.courseId = +route.snapshot.paramMap.get('cid');
     this.examId = route.snapshot.paramMap.get('eid');
+    this.auth.isAdminForCourse(this.courseId).subscribe(is => this.adminAccess = is);
+    this.db.getCourse(this.courseId).subscribe(course => this.course = course);
   }
 
   ngOnInit() {
