@@ -1,6 +1,6 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { QuestionComponent } from './question.component';
+import {QuestionComponent} from './question.component';
 
 describe('QuestionComponent', () => {
   let component: QuestionComponent;
