@@ -1,8 +1,13 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
 import {DbService} from '../../core/db.service';
-import {Question} from '../../core/entities/question';
-import {Solution} from '../../core/entities/solution';
+import {QuestionId} from '../../entities/question';
+import {SolutionId} from '../../entities/solution';
+import {AuthService} from '../../users/auth.service';
+import {flatMap} from 'rxjs/operators';
+import {Observable} from 'rxjs';
+import {TopicWithCreatorId} from '../../entities/topic';
+import {Course} from '../../entities/course';
 
 @Component({
   selector: 'app-question',
@@ -12,25 +17,21 @@ import {Solution} from '../../core/entities/solution';
 export class QuestionComponent implements OnInit {
 
   private id: string;
-  public question: Question;
-  public solutions: Solution[];
+  question$: Observable<QuestionId>;
+  solutions$: Observable<SolutionId[]>;
+  topics$: Observable<TopicWithCreatorId[]>;
+  course$: Observable<Course>;
+  isAdmin$: Observable<boolean>;
 
-  constructor(private route: ActivatedRoute, private db: DbService) {
+  constructor(private route: ActivatedRoute, private db: DbService, private auth: AuthService) {
     this.id = this.route.snapshot.paramMap.get('id');
+    this.topics$ = this.db.getTopicsForQuestion(this.id);
+    this.question$ = this.db.getQuestion(this.id);
+    this.solutions$ = this.db.getSolutions(this.id);
+    this.isAdmin$ = this.db.getQuestion(this.id).pipe(flatMap(q => this.auth.isAdminForCourse(q.course)));
+    this.course$ = this.db.getQuestion(this.id).pipe(flatMap(q => this.db.getCourse(q.course)));
   }
 
   ngOnInit() {
-    this.getQuestion();
-    this.getSolutions();
   }
-
-  getQuestion(): void {
-    this.db.getQuestion(this.id).subscribe(q => this.question = q);
-
-  }
-
-  getSolutions(): void {
-    this.db.getSolutions(this.id).subscribe(sol => this.solutions = sol);
-  }
-
 }
