@@ -1,6 +1,6 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { UserProfileComponent } from './user-profile.component';
+import {UserProfileComponent} from './user-profile.component';
 
 describe('UserProfileComponent', () => {
   let component: UserProfileComponent;
