@@ -1,5 +1,10 @@
-import { Component, OnInit } from '@angular/core';
-import {AuthService} from '../../core/auth.service';
+import {Component, OnInit} from '@angular/core';
+import {AuthService} from '../auth.service';
+import {Observable} from 'rxjs';
+import {UserData} from '../../entities/user';
+import {AngularFirestore} from '@angular/fire/firestore';
+import {ActivatedRoute} from '@angular/router';
+import {DbService} from '../../core/db.service';
 
 @Component({
   selector: 'app-user-profile',
@@ -8,12 +13,15 @@ import {AuthService} from '../../core/auth.service';
 })
 export class UserProfileComponent implements OnInit {
 
-  constructor(public auth: AuthService) { }
+  user$: Observable<UserData>;
+  userId: string;
 
-  ngOnInit() {
+  constructor(public auth: AuthService, private db: DbService, private route: ActivatedRoute) {
+    this.userId = this.route.snapshot.paramMap.get('uid');
+    this.user$ = this.userId ?  this.db.getUser(this.userId) : this.auth.user$;
   }
 
-  logout() {
-    this.auth.signOut();
+  ngOnInit() {
+
   }
 }
