@@ -1,3 +1,11 @@
 export const environment = {
-  production: true
+  production: true,
+  firebase: {
+    apiKey: 'AIzaSyCREv45Llml_gg4hwXeMTXrB6gsDdBkSUQ',
+    authDomain: 'yp5779-testock.firebaseapp.com',
+    databaseURL: 'https://yp5779-testock.firebaseio.com',
+    projectId: 'yp5779-testock',
+    storageBucket: 'yp5779-testock.appspot.com',
+    messagingSenderId: '997206110546'
+  }
 };
